requirements.fastapi.txt
fastapi[all]
uvicorn
pyjwt
asyncpg
psycopg2-binary
aioredis
